backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.13
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
